refactored min-side-laerer as it was unreadable
