+++ enkellaering_ui/app/min-side-laerer/teacherName.tsx
@@ -1,0 +1,28 @@
+"use client"
+import { Teacher } from "../admin/types";
+import { LampContainer } from "@/components/ui/lamp";
+import { motion } from "framer-motion"; 
+
+export function TeacherName({teacher} : {teacher: Teacher}) {
+    const teacherFirstname = teacher.firstname;
+    const teacherLastname = teacher.lastname;
+
+
+
+    return (<>
+        <LampContainer>
+            <motion.h1
+                initial={{ opacity: 0.5, y: 100 }}
+                whileInView={{ opacity: 1, y: 0 }}
+                transition={{
+                delay: 0.3,
+                duration: 0.8,
+                ease: "easeInOut",
+                }}
+                className="mt-8 bg-gradient-to-br from-slate-300 to-slate-500 py-4 bg-clip-text text-center text-4xl font-medium tracking-tight text-transparent md:text-7xl"
+            >
+                {teacherFirstname} {teacherLastname}
+            </motion.h1>
+        </LampContainer>
+    </>)
+}
